+++ include/ifm3d_ros2/ods_module.hpp
@@ -10,9 +10,11 @@
 #include <rclcpp/rclcpp.hpp>
 #include <nav_msgs/msg/occupancy_grid.hpp>
 #include <nav2_msgs/msg/costmap.hpp>
+#include <sensor_msgs/msg/laser_scan.hpp>
 
 #include <ifm3d_ros2/function_module.hpp>
 #include <ifm3d_ros2/msg/zones.hpp>
+#include <ifm3d_ros2/msg/extrinsics_calibration_correction.hpp>
 
 #include <ifm3d/fg/frame.h>
 
@@ -29,6 +31,12 @@ class OdsModule : public FunctionModule, public std::enable_shared_from_this<Ods
   using OccupancyGridMsg = nav_msgs::msg::OccupancyGrid;
   using OccupancyGridPublisher = std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<OccupancyGridMsg>>;
 
+  using PolarOccupancyGridMsg = sensor_msgs::msg::LaserScan;
+  using PolarOccupancyGridPublisher = std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<PolarOccupancyGridMsg>>;
+
+  using ExtrinsicsCalibrationCorrectionMsg = ifm3d_ros2::msg::ExtrinsicsCalibrationCorrection;
+  using ExtrinsicsCalibrationCorrectionPublisher = std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<ExtrinsicsCalibrationCorrectionMsg>>;
+
   using ZonesMsg = ifm3d_ros2::msg::Zones;
   using ZonesPublisher = std::shared_ptr<rclcpp_lifecycle::LifecyclePublisher<ZonesMsg>>;
 
@@ -40,6 +48,8 @@ public:
   nav_msgs::msg::OccupancyGrid extract_ros_occupancy_grid(ifm3d::Frame::Ptr frame);
   nav2_msgs::msg::Costmap extract_ros_costmap(ifm3d::Frame::Ptr frame);
   ifm3d_ros2::msg::Zones extract_zones(ifm3d::Frame::Ptr frame);
+  sensor_msgs::msg::LaserScan extract_ros_polar_occupancy_grid(ifm3d::Frame::Ptr frame);
+  ifm3d_ros2::msg::ExtrinsicsCalibrationCorrection extract_ros_extrinsics_calibration_correction(ifm3d::Frame::Ptr frame);
 
   const std::string get_name()
   {
@@ -64,13 +74,19 @@ private:
   OccupancyGridPublisher ros_occupancy_grid_publisher_;
   CostmapPublisher ros_costmap_publisher_;
   ZonesPublisher zones_publisher_;
+  PolarOccupancyGridPublisher ros_polar_occupancy_grid_publisher_;
+  ExtrinsicsCalibrationCorrectionPublisher ros_extrinsics_calibration_correction_publisher_;
 
   std::string frame_id_;
   bool publish_occupancy_grid_;
   bool publish_costmap_;
+  bool publish_polar_occupancy_grid_;
+  bool publish_extrinsics_calibration_correction_;
   rcl_interfaces::msg::ParameterDescriptor frame_id_descriptor_;
   rcl_interfaces::msg::ParameterDescriptor publish_occupancy_grid_descriptor_;
   rcl_interfaces::msg::ParameterDescriptor publish_costmap_descriptor_;
+  rcl_interfaces::msg::ParameterDescriptor publish_polar_occupancy_grid_descriptor_;
+  rcl_interfaces::msg::ParameterDescriptor publish_extrinsics_calibration_correction_descriptor_;
 };
 
 }  // namespace ifm3d_ros2